# docs/swagger.yaml
basePath: /
definitions:
...
  models.Menu:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      order:
        type: integer
      parent_id:
        type: string
      roles:
        description: '"admin", "editor", "user", "all"'
        items:
          type: string
        type: array
      title:
        type: string
      type:
        description: '"frontend" veya "backend"'
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      url:
        type: string
      visible:
        type: boolean
    required:
    - title
    - type
    type: object
...
  models.Permission:
    properties:
      actions:
        items:
          type: string
        type: array
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      module:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  models.Role:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      is_system:
        type: boolean
      name:
        type: string
      permissions:
        description: Permission referansları
        items:
          type: string
        type: array
      updated_at:
        type: string
      updated_by:
        type: string
    required:
    - name
    type: object
...
info:
  contact:
    email: support@example.com
    name: API Support
  description: This is the API documentation for the admin panel project.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Admin Panel API
  version: "1.0"
paths:
...
  /menus:
    get:
      description: Retrieve all menus, including frontend and backend menus
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Menu type (frontend or backend)
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of menus
          schema:
            items:
              $ref: '#/definitions/models.Menu'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to retrieve menus
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all menus
      tags:
      - Menus
    post:
      consumes:
      - application/json
      description: Add a new menu with its details
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Menu details
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/models.Menu'
      produces:
      - application/json
      responses:
        "201":
          description: Menu created successfully
          schema:
            $ref: '#/definitions/models.Menu'
        "400":
          description: Invalid request payload
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to create menu
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new menu
      tags:
      - Menus
  /menus/{id}:
    delete:
      description: Remove a menu by its unique identifier
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Menu ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid menu ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Menu not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to delete menu
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a menu
      tags:
      - Menus
    put:
      consumes:
      - application/json
      description: Update menu details by its ID
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Menu ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated menu details
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/models.Menu'
      produces:
      - application/json
      responses:
        "200":
          description: Menu updated successfully
          schema:
            $ref: '#/definitions/models.Menu'
        "400":
          description: Invalid request payload or menu ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Menu not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to update menu
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a menu
      tags:
      - Menus
...
  /permissions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Permission'
            type: array
      summary: Get all permission modules
      tags:
      - Permissions
    post:
      consumes:
      - application/json
      parameters:
      - description: Permission module
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Permission'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Create a permission module
      tags:
      - Permissions
  /permissions/{id}:
    delete:
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: string
      responses: {}
      summary: Delete permission module
      tags:
      - Permissions
    get:
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get permission module by ID
      tags:
      - Permissions
    put:
      consumes:
      - application/json
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Permission'
      produces:
      - application/json
      responses: {}
      summary: Update permission module
      tags:
      - Permissions
...
  /roles:
    get:
      description: Retrieve all roles with their permissions and details
      produces:
      - application/json
      responses:
        "200":
          description: List of roles
          schema:
            items:
              $ref: '#/definitions/models.Role'
            type: array
        "500":
          description: Failed to retrieve roles
          schema:
            additionalProperties: true
            type: object
      summary: Get all roles
      tags:
      - Roles
    post:
      consumes:
      - application/json
      description: Add a new role with its permissions and details
      parameters:
      - description: Role details
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/models.Role'
      produces:
      - application/json
      responses:
        "201":
          description: Role created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request payload
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to create role
          schema:
            additionalProperties: true
            type: object
      summary: Create a new role
      tags:
      - Roles
...
security:
- BearerAuth: []
swagger: "2.0"